(hel-741) Retrait du code inutile suite au passage du tri au back
